add data seriealize and pass test

+++ train/src/controller.h
@@ -30,7 +30,7 @@ namespace entity2vec {
         void trainThread(uint32_t threadId);
 
         void train(std::shared_ptr<args> args);
-        void skipgram(model& model, real lr, const std::vector<uint32_t>& line, const std::vector<uint32_t>& labels);
+        void skipgram(model& model, real lr, const std::vector<int64_t>& line, const std::vector<int64_t>& labels);
         void printInfo(real progress, real loss);
         void printWords(std::string word, uint32_t k);
